extract OtlpExportProtocol string extension to separate class
