+++ src/OpenTelemetry.Exporter.OpenTelemetryProtocol/OtlpExporterOptionsHelpers.cs
@@ -1,0 +1,29 @@
+// <copyright file="OtlpExporterOptionsHelpers.cs" company="OpenTelemetry Authors">
+// Copyright The OpenTelemetry Authors
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+// </copyright>
+
+namespace OpenTelemetry.Exporter
+{
+    internal static class OtlpExporterOptionsHelpers
+    {
+        public static OtlpExportProtocol? ToOtlpExportProtocol(this string protocol) =>
+            protocol.Trim() switch
+            {
+                "grpc" => OtlpExportProtocol.Grpc,
+                "http/protobuf" => OtlpExportProtocol.HttpProtobuf,
+                _ => null,
+            };
+    }
+}
