+++ test/OpenTelemetry.Exporter.OpenTelemetryProtocol.Tests/OtlpExporterOptionsHelpersTests.cs
@@ -1,0 +1,34 @@
+// <copyright file="OtlpExporterOptionsHelpersTests.cs" company="OpenTelemetry Authors">
+// Copyright The OpenTelemetry Authors
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+// </copyright>
+
+using Xunit;
+
+namespace OpenTelemetry.Exporter.OpenTelemetryProtocol.Tests
+{
+    public class OtlpExporterOptionsHelpersTests : Http2UnencryptedSupportTests
+    {
+        [Theory]
+        [InlineData("grpc", OtlpExportProtocol.Grpc)]
+        [InlineData("http/protobuf", OtlpExportProtocol.HttpProtobuf)]
+        [InlineData("unsupported", null)]
+        public void ToOtlpExportProtocol_Protocol_MapsToCorrectValue(string protocol, OtlpExportProtocol? expectedExportProtocol)
+        {
+            var exportProtocol = protocol.ToOtlpExportProtocol();
+
+            Assert.Equal(expectedExportProtocol, exportProtocol);
+        }
+    }
+}
